+++ libs/slu-ipc/src/error.rs
@@ -8,6 +8,8 @@ pub enum Error {
     IpcResponseError(String),
     #[error("Serde Json Error: {0}")]
     SerdeJson(#[from] serde_json::Error),
+    #[error("IPC Timeout: {0}")]
+    Timeout(String),
 }
 
 pub type Result<T = ()> = core::result::Result<T, Error>;